// src/app/core/enums/region-code.enum.ts
/** 縣市 碼 */
export enum REGION_CODE {
  /** 全部縣市 */
  ALL = '',

  /** "連江縣" */
  Z = '09007',

  /** "金門縣" */
  W = '09020',

  /** "宜蘭縣" */
  G = '10002',

  /** "新竹縣" */
  J = '10004',

  /** "苗栗縣" */
  K = '10005',

  /** "彰化縣" */
  N = '10007',

  /** "南投縣" */
  M = '10008',

  /** "雲林縣" */
  P = '10009',

  /** "嘉義縣" */
  Q = '10010',

  /** "屏東縣" */
  T = '10013',

  /** "臺東縣" */
  V = '10014',

  /** "花蓮縣" */
  U = '10015',

  /** "澎湖縣" */
  X = '10016',

  /** "基隆市" */
  C = '10017',

  /** "新竹市" */
  O = '10018',

  /** "嘉義市" */
  I = '10020',

  /** "臺北市" */
  A = '63000',

  /** "高雄市" */
  E = '64000',

  /** "新北市" */
  F = '65000',

  /** "臺中市" */
  B = '66000',

  /** "臺南市" */
  D = '67000',

  /** "桃園市" */
  H = '68000',
}